Add an ability to listen for TerminalLine change events
The intended way is to wrap actual line changes into `getLineAndDoWithReportingChanges` extension function call. Then all events emitting logic might be placed right near the actual text entries changes in the TerminalLine.
The alternative approach can be to add the listener to each line and leave it for a whole line lifetime. But it is not really suitable, because it requires tracking of listener presence in the TerminalLine. It is much easier to track actual places where line is changed and specify listening explicitly.